requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
PyInstaller>=5.13.0